+++ es6/prefix-matcher.js
@@ -4,7 +4,8 @@ function replaceNbsps(str) {
 }
 
 function match(condition, placeHolderContent) {
-	if (typeof condition === "string") {
+	const type = typeof condition;
+	if (type === "string") {
 		return (
 			replaceNbsps(placeHolderContent.substr(0, condition.length)) === condition
 		);
@@ -12,18 +13,26 @@ function match(condition, placeHolderContent) {
 	if (condition instanceof RegExp) {
 		return condition.test(replaceNbsps(placeHolderContent));
 	}
+	if (type === "function") {
+		return !!condition(placeHolderContent);
+	}
 }
 function getValue(condition, placeHolderContent) {
-	if (typeof condition === "string") {
+	const type = typeof condition;
+	if (type === "string") {
 		return replaceNbsps(placeHolderContent).substr(condition.length);
 	}
 	if (condition instanceof RegExp) {
 		return replaceNbsps(placeHolderContent).match(condition)[1];
 	}
+	if (type === "function") {
+		return condition(placeHolderContent);
+	}
 }
 
 function getValues(condition, placeHolderContent) {
-	if (typeof condition === "string") {
+	const type = typeof condition;
+	if (type === "string") {
 		return [
 			placeHolderContent,
 			replaceNbsps(placeHolderContent).substr(condition.length),
@@ -31,6 +40,9 @@ function getValues(condition, placeHolderContent) {
 	}
 	if (condition instanceof RegExp) {
 		return replaceNbsps(placeHolderContent).match(condition);
+	}
+	if (type === "function") {
+		return [placeHolderContent, condition(placeHolderContent)];
 	}
 }